MINIFI-226: First Step is to Separate YamlConfiguration from FlowController.
Next step will be to Create Controller services, which will be easier
with the seperated FlowConfiguration
