+++ libminifi/include/FlowController.h
@@ -28,7 +28,6 @@
 #include <atomic>
 #include <algorithm>
 #include <set>
-#include "yaml-cpp/yaml.h"
 
 #include "properties/Configure.h"
 #include "core/Relationship.h"
@@ -39,21 +38,11 @@
 #include "core/ProcessContext.h"
 #include "core/ProcessSession.h"
 #include "core/ProcessGroup.h"
-#include "processors/GenerateFlowFile.h"
-#include "processors/LogAttribute.h"
-#include "processors/RealTimeDataCollector.h"
+#include "core/FlowConfiguration.h"
 #include "TimerDrivenSchedulingAgent.h"
 #include "EventDrivenSchedulingAgent.h"
 #include "FlowControlProtocol.h"
-#include "RemoteProcessorGroupPort.h"
-#include "provenance/Provenance.h"
-#include "processors/GetFile.h"
-#include "processors/PutFile.h"
-#include "processors/TailFile.h"
-#include "processors/ListenSyslog.h"
-#include "processors/ListenHTTP.h"
-#include "processors/ExecuteProcess.h"
-#include "processors/AppendHostInfo.h"
+
 #include "core/Property.h"
 
 namespace org {
@@ -63,22 +52,6 @@ namespace minifi {
 
 // Default NiFi Root Group Name
 #define DEFAULT_ROOT_GROUP_NAME ""
-#define DEFAULT_FLOW_YAML_FILE_NAME "conf/flow.yml"
-#define CONFIG_YAML_PROCESSORS_KEY "Processors"
-
-struct ProcessorConfig {
-  std::string id;
-  std::string name;
-  std::string javaClass;
-  std::string maxConcurrentTasks;
-  std::string schedulingStrategy;
-  std::string schedulingPeriod;
-  std::string penalizationPeriod;
-  std::string yieldPeriod;
-  std::string runDurationNanos;
-  std::vector<std::string> autoTerminatedRelationships;
-  std::vector<core::Property> properties;
-};
 
 /**
  * Flow Controller class. Generally used by FlowController factory
@@ -147,18 +120,6 @@ class FlowController : public core::CoreComponent {
       root_->updatePropertyValue(processorName, propertyName, propertyValue);
   }
 
-  // Create Processor (Node/Input/Output Port) based on the name
-  virtual std::shared_ptr<core::Processor> createProcessor(std::string name,
-                                                           uuid_t uuid) = 0;
-  // Create Root Processor Group
-  virtual core::ProcessGroup *createRootProcessGroup(std::string name,
-                                                     uuid_t uuid) = 0;
-  // Create Remote Processor Group
-  virtual core::ProcessGroup *createRemoteProcessGroup(std::string name,
-                                                       uuid_t uuid) = 0;
-  // Create Connection
-  virtual std::shared_ptr<Connection> createConnection(std::string name,
-                                                       uuid_t uuid) = 0;
   // set 8 bytes SerialNumber
   virtual void setSerialNumber(uint8_t *number) {
     protocol_->setSerialNumber(number);
@@ -171,12 +132,11 @@ class FlowController : public core::CoreComponent {
   // NiFi property File Name
   std::string properties_file_name_;
   // Root Process Group
-  core::ProcessGroup *root_;
+  std::unique_ptr<core::ProcessGroup> root_;
   // MAX Timer Driven Threads
   int max_timer_driven_threads_;
   // MAX Event Driven Threads
   int max_event_driven_threads_;
-  // Config
   // FlowFile Repo
   // Whether it is running
   std::atomic<bool> running_;
@@ -202,9 +162,14 @@ class FlowController : public core::CoreComponent {
   // FlowControl Protocol
   FlowControlProtocol *protocol_;
 
+  // flow configuration object.
+  std::unique_ptr<core::FlowConfiguration> flow_configuration_;
+
   FlowController(std::shared_ptr<core::Repository> provenance_repo,
-                 std::shared_ptr<core::Repository> flow_file_repo)
-      : CoreComponent(core::getClassName<FlowController>()), root_(0),
+                 std::shared_ptr<core::Repository> flow_file_repo,
+                 std::unique_ptr<core::FlowConfiguration> flow_configuration)
+      : CoreComponent(core::getClassName<FlowController>()),
+        root_(nullptr),
         max_timer_driven_threads_(0),
         max_event_driven_threads_(0),
         running_(false),
@@ -213,13 +178,13 @@ class FlowController : public core::CoreComponent {
         flow_file_repo_(flow_file_repo),
         protocol_(0),
         _timerScheduler(provenance_repo_),
-        _eventScheduler(provenance_repo_){
+        _eventScheduler(provenance_repo_),
+        flow_configuration_(std::move(flow_configuration)) {
     if (provenance_repo == nullptr)
       throw std::runtime_error("Provenance Repo should not be null");
     if (flow_file_repo == nullptr)
-          throw std::runtime_error("Flow Repo should not be null");
+      throw std::runtime_error("Flow Repo should not be null");
   }
-
 
 };
 
@@ -255,22 +220,15 @@ class FlowControllerImpl : public FlowController {
       root_->updatePropertyValue(processorName, propertyName, propertyValue);
   }
 
-  // Create Processor (Node/Input/Output Port) based on the name
-  std::shared_ptr<core::Processor> createProcessor(std::string name,
-                                                   uuid_t uuid);
-  // Create Root Processor Group
-  core::ProcessGroup *createRootProcessGroup(std::string name, uuid_t uuid);
-  // Create Remote Processor Group
-  core::ProcessGroup *createRemoteProcessGroup(std::string name, uuid_t uuid);
-  // Create Connection
-  std::shared_ptr<Connection> createConnection(std::string name, uuid_t uuid);
-
   // Constructor
   /*!
    * Create a new Flow Controller
    */
-  FlowControllerImpl(std::shared_ptr<core::Repository> repo,std::shared_ptr<core::Repository> flow_file_repo,
-                     std::string name = DEFAULT_ROOT_GROUP_NAME);
+  FlowControllerImpl(
+      std::shared_ptr<core::Repository> repo,
+      std::shared_ptr<core::Repository> flow_file_repo,
+      std::unique_ptr<core::FlowConfiguration> flow_configuration,
+      std::string name = DEFAULT_ROOT_GROUP_NAME);
 
   // Prevent default copy constructor and assignment operation
   // Only support pass by reference or pointer
@@ -279,29 +237,10 @@ class FlowControllerImpl : public FlowController {
 
  private:
 
-  std::mutex mutex_;
+  std::recursive_mutex mutex_;
 
   // configuration object
   Configure *configure_;
-  // Process Processor Node YAML
-  void parseProcessorNodeYaml(YAML::Node processorNode,
-                              core::ProcessGroup *parent);
-  // Process Port YAML
-  void parsePortYaml(YAML::Node *portNode, core::ProcessGroup *parent,
-                     TransferDirection direction);
-  // Process Root Processor Group YAML
-  void parseRootProcessGroupYaml(YAML::Node rootNode);
-  // Process Property YAML
-  void parseProcessorPropertyYaml(YAML::Node *doc, YAML::Node *node,
-                                  std::shared_ptr<core::Processor> processor);
-  // Process connection YAML
-  void parseConnectionYaml(YAML::Node *node, core::ProcessGroup *parent);
-  // Process Remote Process Group YAML
-  void parseRemoteProcessGroupYaml(YAML::Node *node,
-                                   core::ProcessGroup *parent);
-  // Parse Properties Node YAML for a processor
-  void parsePropertiesNodeYaml(YAML::Node *propertiesNode,
-                               std::shared_ptr<core::Processor> processor);
 
 };
