+++ libminifi/src/Configure.cpp
@@ -1,6 +1,4 @@
 /**
- * @file Configure.cpp
- * Configure class implementation
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
